Assignment six
Create TCP client that writes bytes and reads response
